fix: fixed font data corruption
This was the most annoying bug to find in my life. I was using the
texture() function in my shaders incorrectly. I was supposed to pass in
the Sampler2D (texture slot) but I was just passing in 0 every time. For
some reason all of my tests functioned correctly regardless of this.
Fuck you, OpenGL.